requirements.txt
torch
torchvision
transformers
streamlit
opencv-python
pillow
retina-face
numpy
tf-keras